RGB Settings Menu Fix - Rewrote the menu in order to fix soft lock due to recent changes for View Port Orientation
