+++ applications/settings/cfw_app/cfw_app.h
@@ -42,6 +42,19 @@
 
 ARRAY_DEF(CharList, char*)
 
+static const struct {
+    char* name;
+    RgbColor color;
+} lcd_colors[] = {
+    {"Orange", {255, 60, 0}},  {"Red", {255, 0, 0}},      {"Maroon", {128, 0, 0}},
+    {"Yellow", {255, 150, 0}}, {"Olive", {128, 128, 0}},  {"Lime", {0, 255, 0}},
+    {"Green", {0, 128, 0}},    {"Aqua", {0, 255, 127}},   {"Cyan", {0, 210, 210}},
+    {"Azure", {0, 127, 255}},  {"Teal", {0, 128, 128}},   {"Blue", {0, 0, 255}},
+    {"Navy", {0, 0, 128}},     {"Purple", {128, 0, 255}}, {"Fuchsia", {255, 0, 255}},
+    {"Pink", {255, 0, 127}},   {"Brown", {165, 42, 42}},  {"White", {150, 150, 110}},
+    {"Off", {0, 0, 0}},
+};
+
 typedef struct {
     Gui* gui;
     DialogsApp* dialogs;
