+++ applications/settings/cfw_app/scenes/cfw_app_scene_config.h
@@ -15,5 +15,6 @@ ADD_SCENE(cfw_app, protocols_frequencies_hopper, ProtocolsFrequenciesHopper)
 ADD_SCENE(cfw_app, protocols_frequencies_add, ProtocolsFrequenciesAdd)
 ADD_SCENE(cfw_app, misc, Misc)
 ADD_SCENE(cfw_app, misc_screen, MiscScreen)
+ADD_SCENE(cfw_app, misc_screen_rgb_settings, MiscScreenRGBSettings)
 ADD_SCENE(cfw_app, misc_screen_color, MiscScreenColor)
 ADD_SCENE(cfw_app, misc_rename, MiscRename)
